requirements.txt
amqp==5.1.1
anyio==3.6.2
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
fastapi==0.95.2
h11==0.14.0
idna==3.4
joblib==1.2.0
kombu==5.2.4
numpy==1.24.3
pandas==2.0.1
Pillow==9.5.0
prompt-toolkit==3.0.38
pydantic==1.10.7
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3
redis==4.5.5
requests==2.30.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
-f https://download.pytorch.org/whl/torch_stable.html
torch==1.13.0+cpu
-f https://download.pytorch.org/whl/torch_stable.html
torchvision==0.14.0+cpu
typing_extensions==4.5.0
tzdata==2023.3
urllib3==2.0.2
uvicorn==0.22.0
vine==5.0.0
wcwidth==0.2.6
